requirements.txt
Pillow==6.0.0
cycler==0.10.0
kiwisolve==1.1.0
matplotlib==3.1.1
numpy==1.14.5
opencv-python==3.4.2.16
pandas==0.25.1
pip==19.0.3
pyparsing==2.4.2
python-dateutil==2.8.0
torch==0.3.1
torchvision==0.2.1
scipy==1.3.1
setuptools==40.8.0
six==1.12.0
tqdm==4.38.0
scikit-image==0.15.0
torchtext==0.2.3